Small fixes
Using unitID instead of Unit for removing from militaryUnits map. Added
LinkedList to try to keep track of unit groups, not yet complete.

Issue: Once a unit is being created in the training queue of a barrack
it's considered created by the bwapi. Need a way to make it so that it's
actually "created" when the unit is DONE training.