Created UnitTypesRequest class in MM
I feel like UnitTypes being requested is not being very efficient. i.e.
no flexability in wanting how many of this type of units to go on an
unitOperation instead in the past it just shoves all the first UnitTypes
on the list that are available in a team and then moving on to the next
UnitTypes down the list.

With UnitTypesRequest, you can have a num of units wanted of each
UnitTypes.

Let me know if you think this will help out military orders in CoreBaby
Antonio.