Military teams are updated
When size of a military team is 0 (ALL units are destroyed in the team),
it gets removed. It seems like my method handleTeamStatus was causing a
problem it's commented out for now as well. Things should be working,
let me know if things are normal again Matt.

I left some debug printlns in the console to show it.